Move functions to compute_morphometrics.py
